// app/utils/types.ts
export interface Experience {
    year: string
    title: string
    company: string
    description: string
    color: string
}

export interface Project {
    title: string
    category: string
    description: string
    image: string
    tags: string[]
    color: string
    link: string
}

export interface Testimonial {
    quote: string
    author: string
    role: string
    color: string
}